Add a Python script to run tests.
The test suite is organized as a list of programs and their expected
outputs. The input program is listed in a .input file and the expected
output is listed in the .output file of the same name.

The script compiles each input program, and checks the actual output
against the expected output.